+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 0.17.1 - 2024-02-23
+
 ### Added
 
 - Allow computing energy spectrum from unsmoothed vorticity in Fourier space.